B4: signedness/width type fixes (cycle 1) (#17622)
* MenuUtil: m_battleStateFlag s16 not u16 union member (lhz->lha, 96.28%)

The m_battleStateFlag union member was unsigned short while its siblings
m_cmdLayoutFlag/m_singleMenuMode are signed short. DrawHelpMessageUS reads
it via static_cast<int>(m_battleStateFlag) which the target sign-extends
(lha) but ours zero-extended (lhz). Making it signed short matches the
target's lha and aligns the union member signedness.


* wm_menu: DrawMCList __cntlzw msgId logical shift (srawi->srwi, 95.04%)

The __cntlzw-based msgId computed (__cntlzw(...) >> 5) where the cntlzw
result was treated as signed int, emitting an arithmetic srawi. The cntlzw
result is non-negative and the sibling site at line 4088 already wraps it
in static_cast<unsigned int> before >> 5. Wrapping here matches the target's
logical srwi r19,r0,5.


---------

+++ src/wm_menu.cpp
@@ -11508,7 +11508,7 @@ LAB_next:
 				fontF8->SetColor(CColor(0xFF, 0xFF, 0xFF, 0xFF).color);
 				fontF8->SetTlut(0x19);
 				const unsigned int msgId =
-					__cntlzw(static_cast<unsigned int>(static_cast<int>(*reinterpret_cast<char*>(slotData + 0x42)))) >> 5;
+					static_cast<unsigned int>(__cntlzw(static_cast<unsigned int>(static_cast<int>(*reinterpret_cast<char*>(slotData + 0x42))))) >> 5;
 				const int width = static_cast<int>(fontF8->GetWidth(const_cast<char*>(GetMcStr(msgId))));
 				const double* pHd1 = &DOUBLE_803313F8;
 				const float* pD8f1 = &FLOAT_803314D8;
